Upload, Link User & Video
- Learned how to limit the maximum file size
- Learned how to put the owner's objectId into the video schema
- Learned when we should use populate() or not
- Learned isModified() to check if the value is modified

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
@@ -10,24 +11,38 @@ export const home = async (req, res) => {
 };
 export const see = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (video) {
     return res.render("watch", { pageTitle: video.title, video });
   }
   return res.render("404", { pageTitle: "Video not found" });
 };
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+  } = req;
   const video = await Video.findById(id);
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
   if (!video) {
     return res.render("404", { pageTitle: "Video not found" });
   }
   return res.render("edit", { pageTitle: `Editing`, video });
 };
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
+  const {
+    params: { id },
+    session: { user },
+    body: { title, description, hashtags },
+  } = req;
   const video = await Video.exists({ _id: id });
+  if (String(user._id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
@@ -42,18 +57,29 @@ export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: `Upload Video` });
 };
 export const postUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
+  const {
+    session: {
+      user: { _id },
+    },
+    file,
+    body: { title, description, hashtags },
+  } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      owner: _id,
       createdAt: Date.now(),
+      fileUrl: file.path,
       hashtags: Video.formatHashtags(hashtags),
       meta: {
         view: 0,
         rating: 0,
       },
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect(`/`);
   } catch (error) {
     return res.status(400).render("upload", {
@@ -71,12 +97,23 @@ export const search = async (req, res) => {
   return res.render("search", { pageTitle: "Search", videos });
 };
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
   try {
+    const video = Video.findById(id);
+    if (!video) {
+      return res.status(404).redirect("404", { pageTitle: "Video Not Found" });
+    }
+    if (String(_id) !== String(video.owner)) {
+      return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
   } catch (error) {
     return res.status(404).render("404", { pageTitle: "Video can't delete" });
   }
 };
-export const upload = (req, res) => res.send("Upload Video");
